pyproject.toml
[project]
name = "vertex-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudpickle>=3.1.1",
    "google-adk>=1.14.1",
    "google-genai>=1.36.0",
    "ipykernel>=6.30.1",
    "PyJWT>=2.8.0",
    "pip>=25.2",
    "jwt>=1.4.0",
    "a2a-sdk>=0.3.6",
  
]
